==> 4_fishing_cost_setup.Rmd <==
---
title: "Artisinal Fishing Pressure Cost Set Up"
output: html_document
---
## In this R markdown we will create the first cost layer - artisinal fishing pressure

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# Load packages
library(here)
library(tidyverse)

```

```{r}
## Grab artisinal fishing pressure from local computer 
art_fish_path <- 'G:/group_project/data/fishing_pressure/art_fish_2013_halpern_et_al_2015.tif'

## Read in the Moz EEZ tif created in Script 1
mz_rast_id <- raster(here("rasters/mz_id.tif"))

## Create a raster from fishing pressure layer
art_fish_rast <- raster(art_fish_path)

## Reproject the raster to match the extent of the mz_rast
fishingvalue <- raster::projectRaster(art_fish_rast, mz_rast_id, 
                                            res = 10000,
                                            method = 'ngb')

## Create a data frame from the fishing value raster to see what value range we are working with
fishing_df <- as.data.frame(fishingvalue)

## Priortizr does not like cost values of NA or 0 so changed both to 0.1
fishingvalue[is.na(fishingvalue[])] <- 0.1

## Plot to see what it looks like
plot(fishingvalue)

## Since we changed all NAs values to 0.1 we need to mask this to MZ eez since we are now picking up terrestrial areas 
fishing_eez <- mask(fishingvalue, mz_rast_id)

## Replot to see what it looks like
plot(fishing_eez)

## Looks good lets save as a tif
writeRaster(fishing_eez, here('rasters/fishing.tif'), overwrite = TRUE)

```
