==> 7_model_run_fishing_pressure.Rmd <==
---
title: "Model runs using artisinal fishing pressure as cost"
output: html_document
---

### Now lets to do our next set of runs using artisinal fishing pressure as a cost

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

# Load packages
library(prioritizr)
library(here)
library(tidyverse)
library(gurobi)
library(tmap)

```

## Again, we must assign planning unit, cost, and conservation features

```{r}
## Make sure you have already run script 6 - this will load the PU and Conservation Features in your local environment

## Now we need to load a new cost layer
## Read in  artisanal fisheries raster from tif file
fishing_raster <- raster(here("rasters", "fishing.tif"))

## Create a data frame to check values
fishing_df <- as.data.frame(fishing_raster)

## Let's scale all values 0-1 by dividing by max value using log-transformed +1 fishing values (+1 so that zero values are transformed to zero not NA)
fishing<- log(fishing_raster+1)/cellStats(log(fishing_raster+1), max)

## Create another data frame to make sure it looks good
fishing_df_scaled <- as.data.frame(fishing)

## Plot to make sure it looks good
plot(fishing)

## we are ready to assign cost
cost_fish <- fishing

# Write wraster
writeRaster(cost_fish, here('rasters/fish_cost.tif'), overwrite = TRUE)
```


### Now it's time to start building the problem or simulations for these runs using similar parameters from script 5

## SCENARIO 1: Baseline - 20% target, Boundary Penalty = 0 
```{r}
### OBJECTIVE: Minimize fishing pressure while meeting 20% representation targets, locking in MPAs, locking out oil rigs, and a boundary penalty of 0

## Define the problem with the cost, conservation features, targets, locked in and locked out constraints using the minimum set objective
prob_fish <- problem(cost_fish, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_locked_in_constraints(mpa_raster) %>%
  add_locked_out_constraints(oil_raster) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)
  
## Solve problem
sprob_fish<- solve(prob_fish) 

## Plot the solution to see what it looks like
plot(sprob_fish,  main ="Baseline: Fishing Pressure - 20% Target")

writeRaster(sprob_fish, here('outputs/fish_baseline.tif'), overwrite = TRUE)

```


# SCENARIO 2: SENSITIVITY BOUNDARIES

# Boundary Penalty = 0.0001
```{r}
### Now lets try setting a boundary length of 0.0001
### OBJECTIVE: Minimize fishing pressure while meeting 20% representation targets, locking in MPAs, locking out oil rigs, and a boundary penalty of 0.0001

prob_fish_blm_0001 <- problem(cost_fish, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_locked_in_constraints(mpa_raster) %>%
  add_locked_out_constraints(oil_raster) %>%
  add_boundary_penalties(penalty=0.0001) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

## Solve problem
sprob_fish_blm_0001 <- solve(prob_fish_blm_0001)

## Plot the solution to see what it looks like
plot(sprob_fish_blm_0001,  main = c("BLM = 0.0001"))

writeRaster(sprob_fish_blm_0001, here('outputs/fish_blm_0001.tif'), overwrite = TRUE)

#Found heuristic solution: objective 437.7514884
# Explored 1627 nodes (632174 simplex iterations) in 1200.03 seconds
#Thread count was 1 (of 8 available processors)

#Solution count 10: 233.352 234.358 235.443 ... 242.268

#Time limit reached
#Best objective 2.333523694617e+02, best bound 1.759145645280e+02, gap 24.6142%
```

# Boundary penalty = .001
```{r}
### Now lets try setting a boundary length of .001
### OBJECTIVE: Minimize area while meeting 20% representation targets, locking in MPAs, locking out oil rigs, and a boundary penatly of 1

prob_fish_blm_001 <- problem(cost_fish, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_locked_in_constraints(mpa_raster) %>%
  add_locked_out_constraints(oil_raster) %>%
  add_boundary_penalties(penalty=.001) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

## Solve problem
sprob_fish_blm_001 <- solve(prob_fish_blm_001)

## Plot the solution to see what it looks like
plot(sprob_fish_blm_001 ,  main = c("BLM = .001"))

writeRaster(sprob_fish_blm_0001, here('outputs/fish_blm_001.tif'), overwrite = TRUE)

#Found heuristic solution: objective 4275.5416714
# Explored 1361 nodes (596533 simplex iterations) in 1200.01 seconds
#Thread count was 1 (of 8 available processors)

#Solution count 10: 2049.19 2049.38 2259.68 ... 3070.65

#Time limit reached
#Best objective 2.049188931272e+03, best bound 1.482067286538e+03, gap 27.6754%
```

# Boundary penalty = 1
```{r}
### Now lets try setting a boundary length of 1
### OBJECTIVE: Minimize area while meeting 20% representation targets, locking in MPAs, locking out oil rigs, and a boundary penatly of 1

prob_fish_blm_1 <- problem(cost_fish, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_locked_in_constraints(mpa_raster) %>%
  add_locked_out_constraints(oil_raster) %>%
  add_boundary_penalties(penalty=1) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

## Solve problem
sprob_fish_blm_1 <- solve(prob_fish_blm_1)

## Plot the solution to see what it looks like
plot(sprob_fish_blm_1,  main = c("BLM = 1"))

writeRaster(sprob_fish_blm_1, here('outputs/fish_blm_1.tif'), overwrite = TRUE)

```

# Boundary penalty = 10
```{r}
### Now lets try setting a boundary length of 10
### OBJECTIVE: Minimize area while meeting 20% representation targets, locking in MPAs, locking out oil rigs, and a boundary penatly of 1

prob_fish_blm_10 <- problem(cost_fish, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_locked_in_constraints(mpa_raster) %>%
  add_locked_out_constraints(oil_raster) %>%
  add_boundary_penalties(penalty=10) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

## Print the problem to verify it is as expected   
print(prob_fish_blm_10)

## Solve problem
sprob_fish_blm_10 <- solve(prob_fish_blm_10)

## Plot the solution to see what it looks like
plot(sprob_fish_blm_10,  main = c("BLM = 10"))

```

### SCENARIO 3: No locked in or locked out constraints
```{r}
## SENSITIVTY ANALSYSIS TEST: Now lets run a sensitivity analysis with nothing locked in or out
### OBJECTIVE: Minimize fishing pressure while meeting 20% representation targets, without locking in MPAs, locking out oil rigs, and a boundary penalty of 0

## Establish the problem
prob_fish_no_mpas <- problem(cost_fish, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

#solve problem
sprob_fish_no_mpas <- solve(prob_fish_no_mpas)

plot(sprob_fish_no_mpas,  main = c("Fishing Pressure - no MPAs"))

writeRaster(sprob_fish_no_mpas, here('outputs/fish_mpas.tif'), overwrite = TRUE)

#Found heuristic solution: objective 37.9780379
#Optimal solution found (tolerance 1.00e-01)
#Best objective 2.121596078794e+01, best bound 2.121129170119e+01, gap #0.0220%
```

### SCENARIO 4: Varying Targets

## 10% Targets
```{r}
## SENSITIVTY ANALSYSIS TEST: Now lets run a sensitivity analysis with 10% targets
### OBJECTIVE: Minimize area while meeting 10% representation targets, without locking in MPAs, locking out oil rigs, and a boundary penalty of 0

## Establish the problem
prob_10<- problem(cost_fish, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.1) %>%
  add_locked_in_constraints(mpa_raster) %>%
  add_locked_out_constraints(oil_raster) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

## Solve the problem
sprob_10<- solve(prob_10)


## Plot the problem to see results
plot(sprob_10,  main ="Fishing Pressure - 10% Target w MPAs")

writeRaster(sprob_10, here('outputs/fish_10.tif'), overwrite = TRUE)

```

## 50% Targets
```{r}
## SENSITIVTY ANALSYSIS TEST: Now lets run a sensitivity analysis with 50% targets
### OBJECTIVE: Minimize area while meeting 50% representation targets, without locking in MPAs, locking out oil rigs, and a boundary penatly of 0

## Establish the problem
prob_50<- problem(cost_fish, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.5) %>%
  add_locked_in_constraints(mpa_raster) %>%
  add_locked_out_constraints(oil_raster) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

## Solve the problem
sprob_50<- solve(prob_50)

## Plot to see results
plot(sprob_50,  main ="Fishing Pressure - 50% Target w MPAs")

writeRaster(sprob_50, here('outputs/fish_50.tif'), overwrite = TRUE)

#Solution count 2: 55.7873 81.9724 

#Optimal solution found (tolerance 1.00e-01)
#Best objective 5.578733461421e+01, best bound 5.578390619593e+01, gap 0.0061%

#Found heuristic solution: objective 81.9724143

```

### OPTIONAL SCENARIO 6: 20% with Spearnose skate removed (Rostroraja alba) FIS-131821

```{r}

# Conservation Features without spearnose skate
# Create a raster stack from the multilayer habitat raster and the species distirbution tifs
features_stack_nospear = stack(here("rasters", "multilayer_feature.tif"), here("rasters", "species", "species_Fis-31568.tif"),  here("rasters", "species", "species_Fis-23898.tif"), here("rasters", "species", "species_Fis-23899.tif"), here("rasters", "species", "species_Fis-23061.tif"), here("rasters", "species", "species_Fis-23064.tif"), here("rasters", "species", "species_Fis-58485.tif"), here("rasters", "species", "species_Fis-29423.tif"), here("rasters", "species", "species_Fis-29388.tif"), here("rasters", "species", "species_Fis-30583.tif"), here("rasters", "species", "species_Fis-23273.tif"), here("rasters", "species", "species_Fis-23274.tif"), here("rasters", "species", "species_Fis-23277.tif"), here("rasters", "species", "species_Fis-8339.tif"), here("rasters", "species", "species_Fis-47352.tif"), here("rasters", "species", "species_Fis-24098.tif"), here("rasters", "species", "species_Fis-163295.tif"), here("rasters", "species", "species_Fis-61508.tif"), here("rasters", "species", "species_Fis-30521.tif"), here("rasters", "species", "species_Fis-32599.tif"), here("rasters", "species", "species_Fis-32975.tif"))

## make a data frame of the raster stack to double check everything looks as expected 
con_features_df_nospear <- as.data.frame(features_stack_nospear)

# Establish new problem without spearnose skate
prob_fish_nospear<- problem(cost_fish, features = features_stack_nospear) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_locked_in_constraints(mpa_raster) %>%
  add_locked_out_constraints(oil_raster) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

## Solve the problem
sprob_fish_nospear <- solve(prob_fish_nospear)

## Plot to see results
plot(sprob_fish_nospear,  main ="Fishing Pressure - 20% Target, without spearnose skate")

writeRaster(sprob_fish_nospear, here('outputs/no_spearnose.tif'), overwrite = TRUE)
```


### Now lets plot results next to each other to compare
```{r}
# Now lets make a results graph comparing the different targets 10, 20 and 50
plotstack <- stack(sprob_10, sprob, sprob_50)

tm_shape(plotstack)+
tm_raster(palette = c("#c6c5c5", "#409a00"), n=2, legend.show = FALSE) +
  tm_layout(title = c("(a)", "(b)", "(c)"), panel.labels = c("10%", "20%", "50%"), panel.label.bg.color = "goldenrod3", panel.label.color = 'white', panel.label.size = 1)

tmap_save(filename = "outputs/fishing_varying_targets.png")

```

## Now let's plot boundary penalties, 0.1, 1 and 10 results next to each other to compare
```{r}
plotstack_fish_blm <- stack(sprob_fish_blm, sprob_fish_blm_1, sprob_fish_blm_10)


tm_shape(plotstack_fish_blm)+
  tm_raster(palette = c("#c6c5c5", "#409a00"), n=2, legend.show = FALSE)+
  tm_layout(title = c("(a)", "(b)", "(c)"), panel.labels = c("Penalty = 0.1", "Penatly = 1", "Penatly = 10"), panel.label.bg.color = "darkcyan", panel.label.color = 'white', panel.label.size = 1)

tmap_save(filename = "outputs/fish_blms.png")

```

```{r}
# Now lets make a results graph comparing the MPAS and oil constraints included versues excluded
plotstack <- stack(sprob_fish, sprob_no_mpas)

tm_shape(plotstack)+
tm_raster(palette = c("#c6c5c5", "#409a00"), n=2, legend.show = FALSE) +
  tm_layout(title = c("(a)", "(b)"), panel.labels = c("With Constriants", "Without Constraints"), panel.label.bg.color = "royalblue4", panel.label.color = 'white', panel.label.size = 1)

tmap_save(filename = "outputs/pu_constriants.png")
```
