==> 8_model_run_ports.Rmd <==
---
title: "Model runs using distance to ports  as cost"
output: html_document
---

### Now lets to do our next set of runs using distance to ports  as a cost

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(prioritizr)
library(here)
library(tidyverse)
library(tmap)
```

## Again, we must assign planning unit, cost, and conservation features
```{r}
## Make sure you have already run script 5 - this will load the PU and Conservation Features in your local environment

## Now we need to load a new cost layer
port_raster <- raster(here("rasters", "port_dist_rast.tif"))

## Plot and create data frame to check what it looks like
plot(port_raster)
port_df <- as.data.frame(port_raster)

#scale all values 0-1 by dividing by max value using log-transformed +1 ports values (+1 so that zero values are transformed to zero not NA)
ports<- log(port_raster+1)/ cellStats(log(port_raster+1), max)

## Create anothe data frame to make sure it looks good
ports_df_scaled <- as.data.frame(ports)
plot(ports)

## Assign the new ports raster as costs and call it "costs_ports"
cost_ports <- ports

writeRaster(cost_ports, here('rasters/ports_cost.tif'), overwrite = TRUE)
```

### Now it's time to start building the problem or simulations for these runs using similar parameters from script 5 & 6

## SCENARIO 1: Baseline - 20% target, Boundary Penalty = 0 
```{r}
### OBJECTIVE: Minimize fishing pressure while meeting 20% representation targets, locking in MPAs, locking out oil rigs, and a boundary penalty of 0

## Define the problem with the cost, conservation features, targets, locked in and locked out constraints using the minimum set objective

prob_ports <- problem(cost_ports, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_locked_in_constraints(mpa_raster) %>% 
  add_locked_out_constraints(oil_raster) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)


#solve problem
sprob_ports <- solve(prob_ports)

#plot to see results
plot(sprob_ports,  main = c("Baseline: Ports - 20% Targets"))

writeRaster(sprob_ports, here('outputs/ports_baseline.tif'), overwrite = TRUE) 

#Best objective 1.049229181250e+03, best bound 1.048713945055e+03, gap 0.0491%
```

## SCENARIO 2: TEST BOUNDARIES

# Boundary Penalty = 0.001
```{r}
### Now lets try setting a boundary length of 0..001
### OBJECTIVE: Minimize distance to ports cost  while meeting 20% representation targets, locking in MPAs, locking out oil rigs, and a boundary penalty of 0.5

prob_ports_blm <- problem(cost_ports, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_locked_in_constraints(mpa_raster) %>%
  add_locked_out_constraints(oil_raster) %>%
  add_boundary_penalties(penalty=0.001) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

## Print the problem to verify it is as expected   
print(prob_ports_blm )

## Solve problem
sprob_ports_blm <- solve(prob_ports_blm )

## Plot the solution to see what it looks like
plot(sprob_ports_blm,  main = c("Ports: BLM = 0.001"))

writeRaster(sprob_ports_blm, here('outputs/ports_001.tif'), overwrite = TRUE)
```

## Boundary penalty = 1
```{r}
### Now lets try setting a boundary length of 1
### OBJECTIVE: Minimize area while meeting 20% representation targets, locking in MPAs, locking out oil rigs, and a boundary penatly of 1

prob_ports_blm_1 <- problem(cost_ports, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_locked_in_constraints(mpa_raster) %>%
  add_locked_out_constraints(oil_raster) %>%
  add_boundary_penalties(penalty=1) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

## Print the problem to verify it is as expected   
print(prob_ports_blm_1)

## Solve problem
sprob_ports_blm_1 <- solve(prob_ports_blm_1)

## Plot the solution to see what it looks like
plot(sprob_ports_blm_1,  main = c("BLM = 1"))

writeRaster(sprob_ports_blm_1, here('outputs/ports_1.tif'), overwrite = TRUE)

```

## Boundary penalty = 10
```{r}
### Now lets try setting a boundary length of 10
### OBJECTIVE: Minimize area while meeting 20% representation targets, locking in MPAs, locking out oil rigs, and a boundary penatly of 1

prob_ports_blm_10 <- problem(cost_ports, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_locked_in_constraints(mpa_raster) %>%
  add_locked_out_constraints(oil_raster) %>%
  add_boundary_penalties(penalty=1) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

## Print the problem to verify it is as expected   
print(prob_ports_blm_10)

## Solve problem
sprob_ports_blm_10 <- solve(prob_ports_blm_10)

## Plot the solution to see what it looks like
plot(sprob_ports_blm_10,  main = c("BLM = 10"))

```

### SCENARIO 3: No locked in or locked out constraints
```{r}
## SENSITIVTY ANALSYSIS TEST: Now lets run a sensitivity analysis with nothing locked in or out
### OBJECTIVE: Minimize distance to ports  while meeting 20% representation targets, without locking in MPAs, locking out oil rigs, and a boundary penalty of 0

prob_ports_no_mpas <- problem(cost_ports, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

print(prob_ports_no_mpas)

#solve problem
sprob_ports_no_mpas <- solve(prob_ports_no_mpas)

plot(sprob_ports_no_mpas,  main = c("Fishing Pressure - no MPAs"))

writeRaster(sprob_ports_no_mpas, here('outputs/ports_mpas.tif'), overwrite = TRUE)

#Optimal solution found (tolerance 1.00e-01)
#Best objective 1.046915289600e+03, best bound 1.046556968228e+03, gap #0.0342%
## Now we can compare this result to the baseline result found in Scenario 1
```

### SCENARIO 4: Varying Targets
```{r}
## SENSITIVTY ANALSYSIS TEST: Now lets run a sensitivity analysis with 10% targets
### OBJECTIVE: Minimize area while meeting 10% representation targets, without locking in MPAs, locking out oil rigs, and a boundary penalty of 0

ports_10 <- problem(cost_ports, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.1) %>% 
  add_locked_in_constraints(mpa_raster) %>% 
  add_locked_out_constraints(oil_raster) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)
  
# solve problem
sprob_ports_10 <- solve(ports_10)

# plot resutls
plot(sprob_ports_10,  main = c("Ports - 10% Targets"))

writeRaster(sprob_ports_10, here('outputs/ports_10.tif'), overwrite = TRUE)
```

# SCENARIO 5: 50% Targets
```{r}
## SENSITIVTY ANALSYSIS TEST: Now lets run a sensitivity analysis with 50% targets
### OBJECTIVE: Minimize area while meeting 50% representation targets, without locking in MPAs, locking out oil rigs, and a boundary penatly of 0

ports_50 <- problem(cost_ports, features = features_stack) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.5) %>% 
  add_locked_in_constraints(mpa_raster) %>% 
  add_locked_out_constraints(oil_raster) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)
  
# solve problem
sprob_ports_50 <- solve(ports_50)

# print results
plot(sprob_ports_50,  main = c("Ports - 50% Targets"))

writeRaster(sprob_ports_50, here('outputs/ports_50.tif'), overwrite = TRUE)

```

### OPTIONAL - SCENARIO 6: 20% with Spearnose skate removed (Rostroraja alba) FIS-131821

```{r}

# Conservation Features without spearnose skate
# Create a raster stack from the multilayer habitat raster and the species distirbution tifs
features_stack_nospear = stack(here("rasters", "multilayer_feature.tif"), here("rasters", "species", "species_Fis-31568.tif"),  here("rasters", "species", "species_Fis-23898.tif"), here("rasters", "species", "species_Fis-23899.tif"), here("rasters", "species", "species_Fis-23061.tif"), here("rasters", "species", "species_Fis-23064.tif"), here("rasters", "species", "species_Fis-58485.tif"), here("rasters", "species", "species_Fis-29423.tif"), here("rasters", "species", "species_Fis-29388.tif"), here("rasters", "species", "species_Fis-30583.tif"), here("rasters", "species", "species_Fis-23273.tif"), here("rasters", "species", "species_Fis-23274.tif"), here("rasters", "species", "species_Fis-23277.tif"), here("rasters", "species", "species_Fis-8339.tif"), here("rasters", "species", "species_Fis-47352.tif"), here("rasters", "species", "species_Fis-24098.tif"), here("rasters", "species", "species_Fis-163295.tif"), here("rasters", "species", "species_Fis-61508.tif"), here("rasters", "species", "species_Fis-30521.tif"), here("rasters", "species", "species_Fis-32599.tif"), here("rasters", "species", "species_Fis-32975.tif"))

## make a data frame of the raster stack to double check everything looks as expected 
con_features_df_nospear <- as.data.frame(features_stack_nospear)

# Establish new problem without spearnose skate
prob_ports_nospear<- problem(cost_fish, features = features_stack_nospear) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_locked_in_constraints(mpa_raster) %>%
  add_locked_out_constraints(oil_raster) %>%
  add_gurobi_solver(gap = 0.1, time_limit = 1200)

## Solve the problem
sprob_ports_nospear <- solve(prob_ports_nospear)

## Plot to see results
plot(sprob_ports_nospear,  main ="Distance to Ports - 20% Target, without spearnose skate")

```


### Now lets plot 10%, 20% and 50% target results next to each other to compare
```{r}
plotstack_targets <- stack(sprob_ports_10, sprob_ports, sprob_ports_50)

tm_shape(plotstack_targets)+
tm_raster(palette = c("#c6c5c5", "#409a00"), n=2, legend.show = FALSE) +
  tm_layout(title = c("(a)", "(b)", "(c)"))

tmap_save(filename = "outputs/ports_varying_targets.png")

```

## Now let's plot boundary penalties, 0.5, 1 and 10 results next to each other to compare
```{r}
plotstack_port_blm <- stack(sprob_ports_blm, sprob_ports_blm_1, sprob_ports_blm_10)

tm_shape(plotstack_port_blm)+
tm_raster(palette = c("#c6c5c5", "#409a00"), n=2, legend.show = FALSE) +
  tm_layout(title = c("(a)", "(b)", "(c)"))

tmap_save(filename = "outputs/ports_varying_targets.png")

```

## Now lets plot baseline restuls of 20% from all three scenarios to compare
```{r}

plotstack_comp <- stack(sprob_area, sprob_fish, sprob_ports)

tm_shape(plotstack_comp)+
tm_raster(palette = c("#c6c5c5", "#409a00"), n=2, legend.show = FALSE)+
  tm_layout(title = c("(a)", "(b)", "(c)"), panel.labels = c("Area", "Artisanal Fishing", "Distance to Ports"), panel.label.bg.color = "lightblue4", panel.label.color = 'white', panel.label.size = 1)

tmap_save(filename = "outputs/comparison_all.png")
```

# Now let's plot the baseline results 20% from all three scenarios at BLMs = 0.5
```{r}

plotstack_all_blm <- stack(sprob_area_blm, sprob_fish_blm, sprob_ports_blm)

tm_shape(plotstack_all_blm)+
tm_raster(palette = c("#c6c5c5", "#409a00"), n=2, legend.show = FALSE) +
  tm_layout(title = c("(a)", "(b)", "(c)"))

tmap_save(filename = "outputs/comparison_20_blm.png")

```
